// packages/connect/ic/providers/types.ts
export enum CreateActorError {
    FetchRootKeyFailed = 'FETCH_ROOT_KEY_FAILED',
    CreateActorFailed = 'CREATE_ACTOR_FAILED',
    NotInitialized = 'NOT_INITIALIZED',
    LocalActorsNotSupported = 'LOCAL_ACTORS_NOT_SUPPORTED',
}

export type CreateActorResult<Service> = Result<
    ActorSubclass<Service>,
    CustomError<CreateActorError>
>;

export enum ConnectError {
    NotInitialized = 'NOT_INITIALIZED',
    NotInstalled = 'NOT_INSTALLED',
    ConnectFailed = 'CONNECT_FAILED',
    IsLocked = 'IS_LOCKED',
}

export type ConnectResult = Result<boolean, CustomError<ConnectError>>;

export enum DisconnectError {
    DisconnectFailed = 'DISCONNECT_FAILED',
    NotInitialized = 'NOT_INITIALIZED',
}

export type DisconnectResult = Result<boolean, CustomError<DisconnectError>>;

export enum InitError {
    NotInstalled = 'NOT_INSTALLED',
    InitFailed = 'INIT_FAILED',
    FetchRootKeyFailed = 'FETCH_ROOT_KEY_FAILED',
}
// ...
export enum BalanceError {
    NotInitialized = 'NOT_INITIALIZED',
    QueryBalanceFailed = 'QUERY_BALANCE_FAILED',
}
// ...
export enum TransferError {
    InsufficientBalance = 'INSUFFICIENT_BALANCE',
    TransferFailed = 'TRANSFER_FAILED',
    FaultyAddress = 'FAULTY_ADDRESS',
    NotInitialized = 'NOT_INITIALIZED',
    TokenNotSupported = 'TOKEN_NOT_SUPPORTED',
    NotConnected = 'NOT_CONNECTED',
}